LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-key-operations-expected.txt

Tests if calendar picker key bindings work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
Check that page popup exists.
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
PASS isCalendarTableScrollingWithAnimation() is false
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "2000-01-02"
PASS highlightedDayCells() is "2000-01-02"
Check that arrow keys work properly even when going between weeks and months.
press up
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press up
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-19"
PASS highlightedDayCells() is "1999-12-19"
press down
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press down
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "2000-01-02"
PASS highlightedDayCells() is "2000-01-02"
press up
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press left
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-25"
PASS highlightedDayCells() is "1999-12-25"
press right
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press pageDown
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press pageUp
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press m
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "2000-01-23"
PASS highlightedDayCells() is "2000-01-23"
press shift + m
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press y
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2000-12-24"
PASS highlightedDayCells() is "2000-12-24"
press shift + y
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press d
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2009-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2009-12-27"
PASS highlightedDayCells() is "2009-12-27"
press shift + d
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
press enter
PASS currentMonth() is "1999-12"
PASS selectedValue() is "1999-12-26"
PASS selectedDayCells() is "1999-12-26"
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is "1999-12-26"
PASS document.getElementById("date").value is "1999-12-26"
Check if tabbing works.
PASS focusedElement() is ".list-view calendar-table-view"
PASS focusedElement() is ".calendar-navigation-button"
PASS focusedElement() is ".calendar-navigation-button today-button"
PASS focusedElement() is ".calendar-navigation-button"
PASS focusedElement() is ".month-popup-button"
open the month popup
PASS popupWindow.global.picker.monthPopupView.isVisible is true
PASS highlightedMonthButton() is "1999-12"
press down
PASS highlightedMonthButton() is "2000-04"
press right
PASS highlightedMonthButton() is "2000-05"
press right
PASS highlightedMonthButton() is "2000-06"
press right
PASS highlightedMonthButton() is "2000-07"
press right
PASS highlightedMonthButton() is "2000-08"
press right
PASS highlightedMonthButton() is "2000-09"
press left
PASS highlightedMonthButton() is "2000-08"
press up
PASS highlightedMonthButton() is "2000-04"
press pageDown
PASS highlightedMonthButton() is "2001-04"
press pageUp
PASS highlightedMonthButton() is "2000-04"
select month in the month popup
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS currentMonth() is "2000-04"
PASS selectedValue() is "1999-12-26"
PASS selectedDayCells() is ""
PASS highlightedValue() is "1999-12-26"
PASS highlightedDayCells() is ""
focus calendar table
press t
PASS isCalendarTableScrollingWithAnimation() is true
PASS testToday() || testToday() is true
press esc
PASS document.getElementById("mock-page-popup") is null
PASS successfullyParsed is true

TEST COMPLETE
